Read me's and folder restructuring
