+++ Data/u_values.csv
@@ -1,5 +1,5 @@
 variable,u_value
-firstnm,0.006666666666666667
-lastnm,0.002232142857142857
-sex,0.5555555555555556
-month,0.07905138339920949
+firstnm,0.006696428571428571
+lastnm,0.006666666666666667
+sex,0.5893617021276596
+month,0.08712121212121213
