+++ README.md
@@ -1,4 +1,25 @@
-# ONS Data Linkage Example Scripts
+# ONS Record Linkage Pipeline Example Scripts
+A set of Python scripts to demonstrate deterministic and probabilistic data linkage, using synthetic  census and post-enumeration survey (PES) data.
+
+For additional information consider reading:
+* [Expectation maximisation algorithm](https://courses.cs.washington.edu/courses/cse590q/04au/papers/WinklerEM.pdf) developed by Fellegi-Sunter
+* [ONS data linkage working paper](https://www.ons.gov.uk/methodology/methodologicalpublications/generalmethodology/onsworkingpaperseries/developingstandardtoolsfordatalinkagefebruary2021)
 
 
- todo
+
+## Directories and files of note
+
+Descriptions of project directories and other significant files:
+* `Data/` - the required mock data to operate the scripts at a basic level
+* `lib/` - contains helper functions and configurable parameters
+* `pipeline/` - scripts forming the example record linkage pipeline
+* `setup/` - required packages and versions with instructions on terminal installation via `pip` 
+* `training/` - code to demonstrate key data linkage operations including expectation maximisation
+ 
+
+
+
+
+
+
+
