+++ lib/PARAMETERS.py
@@ -1,6 +1,7 @@
 import os
 
-os.chdir(os.path.pardir)
+# this mess of path finding should always set the working directory to the parent (project) folder
+os.chdir(os.path.join(os.path.join(__file__, os.pardir), os.pardir))
 
 ROOT_DIR = os.getcwd()
 
@@ -8,3 +9,11 @@ DATA_PATH = "Data/"
 
 LIB_PATH = "lib/"
 
+# TODO add in column choice for census and pes as two variables
+
+# TODO add file paths/names for census/pds
+
+CENSUS_FILE_NAME = "Mock_Rwanda_Data_Census.csv"
+PES_FILE_NAME = "Mock_Rwanda_Data_Pes.csv"
+CENSUS_FILE_PATH = DATA_PATH + CENSUS_FILE_NAME
+PES_FILE_PATH = DATA_PATH + PES_FILE_NAME
