+++ pipeline/run_pipeline.py
@@ -1,0 +1,12 @@
+import sys
+sys.path.insert(0, "../")
+from lib.PARAMETERS import *
+# from pipeline import *
+
+
+def run_pipeline():
+    pass
+
+
+if __name__ == "__main__":
+    run_pipeline()
