+++ setup/README.md
@@ -1,8 +1,12 @@
 # Setup Guide
 
-To install the required packages in `requirements.txt` run the following line in terminal:
 
-`pip install -r requirements.txt`
+
+
+
+## Installing the required packages
+To install the required packages in `requirements.txt` run the following line in terminal after navigating to the `setup/` directory:
+* `pip install -r requirements.txt`
