+++ training/Alphaname.py
@@ -2,17 +2,17 @@
 import pandas as pd
 import numpy as np
 import os
-from PARAMETERS import *
+import sys
 
-# Cluster Function
-
+sys.path.insert(0, "../")
+from lib.PARAMETERS import *
 
 # Read in the census data
-CEN = pd.read_csv(DATA_PATH + 'census_cleaned.csv', index_col=False)
-print("Census read in") 
+CEN = pd.read_csv(CENSUS_FILE_PATH, index_col=False)
+print("Census read in")
 
 # Read in the PES data
-PES = pd.read_csv(DATA_PATH + 'pes_cleaned.csv', index_col=False)  
+PES = pd.read_csv(PES_FILE_PATH, index_col=False)
 print("PES read in")
 
 # ------------------------------------------------------------------- #
@@ -20,14 +20,16 @@ print("PES read in")
 # ------------------------------------------------------------------- #    
 
 # Firsly, we need to replace missing names with empty spaces
-CEN['first_name_temporary_cen'] = CEN['first_name_cen'].fillna('')
-CEN['last_name_temporary_cen'] = CEN['last_name_cen'].fillna('')
-PES['first_name_temporary_pes'] = PES['first_name_pes'].fillna('')
-PES['last_name_temporary_pes'] = PES['last_name_pes'].fillna('')
+CEN['first_name_temporary_cen'] = CEN['firstnm_cen'].fillna('')
+CEN['last_name_temporary_cen'] = CEN['lastnm_cen'].fillna('')
+PES['first_name_temporary_pes'] = PES['firstnm_pes'].fillna('')
+PES['last_name_temporary_pes'] = PES['lastnm_pes'].fillna('')
 
 # Now we can create fullname (no spaces) using the temporary name variables
-CEN['fullname_ns_cen'] = CEN[['first_name_temporary_cen', 'last_name_temporary_cen']].apply(lambda row: ''.join(row.values.astype(str)), axis=1)
-PES['fullname_ns_pes'] = PES[['first_name_temporary_pes', 'last_name_temporary_pes']].apply(lambda row: ''.join(row.values.astype(str)), axis=1)
+CEN['fullname_ns_cen'] = CEN[['first_name_temporary_cen', 'last_name_temporary_cen']].apply(
+                            lambda row: ''.join(row.values.astype(str)), axis=1)
+PES['fullname_ns_pes'] = PES[['first_name_temporary_pes', 'last_name_temporary_pes']].apply(
+                            lambda row: ''.join(row.values.astype(str)), axis=1)
 
 # Create alphaname by sorting alphabetically
 CEN['alpha_cen'] = [''.join(sorted(x)) for x in CEN['fullname_ns_cen']]
@@ -40,5 +42,7 @@ PES['alpha_pes'] = np.where(PES['alpha_pes'] == '', None, PES['alpha_pes'])
 # Drop temporary columns no longer needed
 CEN = CEN.drop(['first_name_temporary_cen', 'last_name_temporary_cen', 'fullname_ns_cen'], axis=1)
 PES = PES.drop(['first_name_temporary_pes', 'last_name_temporary_pes', 'fullname_ns_pes'], axis=1)
+
+print(CEN[['firstnm_cen', 'lastnm_cen', 'alpha_cen']].head())
 
 """This code can be added to your cleaning scripts"""
